+++ cotizaciones/src/main/java/com/grupodisatel/cotizaciones/RESTController/RoleController.java
@@ -2,6 +2,8 @@ package com.grupodisatel.cotizaciones.RESTController;
 
 import com.grupodisatel.cotizaciones.Dao.RoleDAO;
 import com.grupodisatel.cotizaciones.Model.Role;
+import com.grupodisatel.cotizaciones.Utils.JWTUtil;
+import com.grupodisatel.cotizaciones.Validation.UserRoleValidation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,19 +16,31 @@ public class RoleController {
     @Autowired
     private RoleDAO roleDAO;
 
-    @RequestMapping(value = "", method = RequestMethod.GET)
-    public List<Role> get(){
-        return roleDAO.getRoles();
+    @Autowired
+    private UserRoleValidation userRoleValidation;
+
+    @GetMapping(value = "")
+    public List<Role> get(@RequestHeader (value = "Authorization") String token){
+        if (userRoleValidation.validatePermission(token, 3)){
+            return roleDAO.getRoles();
+        }
+        return null;
     }
 
     @RequestMapping(value = "getRoleName/{id}", method = RequestMethod.GET)
-    public Role getRoleName(@PathVariable int id){
-        return roleDAO.existRole(id);
+    public Role getRoleName(@PathVariable int id, @RequestHeader(value = "Authorization") String token){
+        if (userRoleValidation.validatePermission(token, 3)){
+            return roleDAO.existRole(id);
+        }
+        return null;
     }
 
     @RequestMapping(value = "delete/{id}", method = RequestMethod.DELETE)
-    public boolean deleteRole(@PathVariable int id){
-        return roleDAO.deleteRole(id);
+    public boolean deleteRole(@PathVariable int id, @RequestHeader(value = "Authorization") String token){
+        if (userRoleValidation.validatePermission(token, 3)){
+            return roleDAO.deleteRole(id);
+        }
+        return false;
     }
 
     @RequestMapping(value = "update/{id}", method = RequestMethod.PUT)
